+++ my-app/public/manifest.json
