+++ my-app/src/thingstodo/ThingsToDo.js
@@ -1,10 +1,9 @@
 import React from 'react';
-import './LookingForMore.css';
 
-function LookingForMore() {
+function ThingsToDo() {
   return (
     <div>
-      <h1>Looking for more ?</h1>
+      <h1>Things to Do</h1>
       <p>Here are some things you can do:</p>
       <ul>
         <li><a href="#">Activity 1</a></li>
@@ -15,6 +14,6 @@ function LookingForMore() {
   );
 }
 
-export default LookingForMore;
+export default ThingsToDo;
 
-
+// Create similar components for Parks and Looking for More
