--- src/thingstodo/ThingsToDo.css
@@ -1,11 +1,0 @@
-body {
-	background: linear-gradient(to bottom, #cef381, #f33369, #5b8060, #e576cb, #eeb013);
-	animation: bg-animation 10s ease infinite;
-};
-
-@keyframes bg-animation {
-	0% { background-position: 0 0; }
-	100% { background-position: 0 100%; }
-};
-
-
